requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
argon2-cffi>=23.1.0
passlib>=1.7.4
python-jose[cryptography]>=3.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
slowapi>=0.1.8
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
aiosqlite>=0.19.0